Added raw masks for Path, Cover & Overlap

+++ JERC/Constants/Colours.cs
@@ -12,7 +12,7 @@ namespace JERC.Constants
     {
         public static Color ColourError => Color.DarkRed;
 
-        public static Color GetGreyscaleGradient(int value) => Color.FromArgb(255, ClampGradientValue(value), ClampGradientValue(value), ClampGradientValue(value));
+        public static Color GetGreyscaleGradient(float value) => Color.FromArgb(255, ClampGradientValue(value), ClampGradientValue(value), ClampGradientValue(value));
 
         public static Color ColourRemove() => Color.Transparent;
         public static Color ColourBrush(Color rgbLow, Color rgbHigh, float percentageAboveMin, int? alphaOverride = null) =>
